Registration page division into parts, image upload component improved, input component made
