+++ pages/mentorRegister.page.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
-import MentorForm from "../components/MentorFormRegistration";
+import MentorForm from "../components/MentorRegistration/MentorFormRegistration";
 import Head from "next/head";
 
 const MentorRegister = () => {
